requirements.txt
tornado
jinja2
PyYAML